refactor(server): extract RouteAssembler from BgpSession — Phase 2 of #93
The per-peer outbound route policy ("which prefixes does this peer get" — the
RU-default vs subscription vs custom-prefix vs custom-AS vs user-source
decision tree, ~234 lines) is extracted into a new RouteAssembler class.
BgpSession.SendAllRoutesAsync becomes a 4-line wrapper that delegates to
RouteAssembler.BuildOutboundRoutesAsync, then does the aggregate + batch +
send (which stays in the session — it owns the _advertisedPrefixes mirror
and the transport).

RouteAssembler absorbs:
- BuildOutboundRoutesAsync (was SendAllRoutesAsync decision tree)
- MakeRoute (route factory)
- AddUserSourceRoutesAsync (user URL source fetch + community stamp)
- GroupByCommunitySet / SameCommunitySet / PartitionByCommunitySet (batch
  partitioning for per-community UPDATEs)

BgpSession keeps: the FSM, transport, timers, teardown latch, send-lock,
advertised-prefixes mirror + lock, RefreshRoutesAsync orchestration, and the
codec glue (SendRoutesAsync/SendRouteBatchAsync/SendUpdateBatchAsync which
calls UpdateCodec from Phase 1).

11 unit tests repointed (BgpSession.X → RouteAssembler.X). All 14 socket
tests stay green as a regression guard. The decision tree logic is unchanged
— only relocated behind the RouteAssembler boundary.
